Separate Redux interfaces
