+++ client/src/redux/interface/notify.ts
@@ -1,0 +1,11 @@
+export interface Notification {
+    type: string;
+    title: string;
+    message: string;
+    dismissable: boolean;
+    duration: number;
+}
+
+export interface Notify {
+    notifications: Notification[];
+}
