+++ client/src/redux/interface/project.ts
@@ -1,0 +1,17 @@
+export interface Project {
+    id: number;
+    title: string;
+    userGroup: number;
+    createdAt: string;
+    modifiedAt: string;
+    createdBy: number;
+    modifiedBy: number;
+    createdByName: string;
+    modifiedByname: string;
+    description: string;
+}
+
+export interface SetProjectAction {
+    userId?: number;
+    project: Project;
+}
