+++ client/src/redux/interface/route.ts
@@ -1,0 +1,11 @@
+export interface Route {
+    path?: string;
+    url?: string;
+    isExact?: boolean;
+    params?: {
+        userId?: number;
+    };
+    routeState?: object; // stricter rule here
+}
+
+
