+++ client/src/redux/interface/slot.ts
@@ -1,0 +1,9 @@
+export interface SlotData {
+    id: number;
+    date: string;
+    startTime: string;
+    endTime: string;
+    user: number;
+    task: number;
+    remarks: string;
+}
