+++ client/src/redux/interface/task.ts
@@ -1,0 +1,14 @@
+export interface Task {
+    id: number;
+    createdAt: string;
+    modifiedAt: string;
+    createdBy: number;
+    modifiedBy: number;
+    createdByName: string;
+    modifiedByName: string;
+    title: string;
+    description: string;
+    project: number;
+}
+
+
