+++ client/src/redux/interface/userGroup.ts
@@ -1,0 +1,9 @@
+export interface UserGroup {
+    id: number;
+    title: string;
+}
+
+export interface SetUserGroupAction {
+    userId?: number;
+    userGroup: UserGroup;
+}
